refactor: Move logic hooks into hooks/logic directory
